Move has_children() to ark
